Commit for backend requests
I have written code to receive and respond properly to creating an account (including the routing to '/{userID}' ). Final testing needs to be done as my backend only responds with errors (has been discussed with Jacob).
Also, currently if creating the account fails, it only sets the error state to "Failed Create Account" as I am unsure of what happens if it errors.